Fix. 작성/수정 페이지 새로고침 오류 동행 신청 및 모집 완료시 수정 페이지 접근 방지 코드 모집완료 뱃지 css 수정
