+++ src/components/partner/partnerUpdateTemplate/PartnerUpdateTemplate.tsx
@@ -2,7 +2,7 @@ import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router';
 import { getInterests } from '../../../api/supabase/interest';
-import { getApplicantList, getConfirmedApplicantList, getPartnerPost, updatePartnerPost } from '../../../api/supabase/partner';
+import { getPartnerPost, updatePartnerPost } from '../../../api/supabase/partner';
 import { Tables } from '../../../api/supabase/supabase';
 import { BtnStyleType } from '../../../types/styleTypes';
 import useSessionStore from '../../../zustand/store';
@@ -16,8 +16,6 @@ import * as St from './style';
 
 function PartnerUpdateTemplate({ postId }: { postId: string }) {
   const { data: partnerPost, isLoading, isError } = useQuery(['partnerPost', postId], () => getPartnerPost({ postId }));
-  const [applicantList, setApplicantList] = useState<Tables<'applicants'>[]>([]);
-  const [confirmedApplicantList, setConfirmedApplicantList] = useState<Tables<'applicants'>[]>([]);
   const [location, setLocation] = useState<string[]>([]);
   const [partnerDates, setPartnerDates] = useState<string[]>([]);
   const [partner, setPartner] = useState<number>(1);
@@ -46,28 +44,6 @@ function PartnerUpdateTemplate({ postId }: { postId: string }) {
   useEffect(() => {
     getInterestsList();
   }, []);
-
-  useEffect(() => {
-    const fetchApplicant = async () => {
-      if (!postId) return;
-      const { data, error } = await getApplicantList(postId);
-      if (error || !data) {
-        setApplicantList([]);
-      } else {
-        setApplicantList(data);
-      }
-    };
-    fetchApplicant();
-    const fetchConfirmedPartnerList = async () => {
-      if (postId) {
-        const response = await getConfirmedApplicantList(postId!);
-        if (response.data !== null) {
-          setConfirmedApplicantList(response.data);
-        }
-      }
-    };
-    fetchConfirmedPartnerList();
-  }, [postId]);
 
   useEffect(() => {
     const fetchPostData = async () => {
@@ -137,11 +113,6 @@ function PartnerUpdateTemplate({ postId }: { postId: string }) {
       navigate(`/partner/detail/${postId}`);
       return false;
     }
-    if (applicantList.length >= 1 || confirmedApplicantList.length >= 1) {
-      AlertError({ title: '수정이 불가능 합니다.', text: '동행 신청이 시작되었습니다.' });
-      navigate(`/partner/detail/${postId}`);
-      return false;
-    }
     if (location.length < 1) {
       AlertWarning({ title: '국가를 선택해주세요.' });
       return false;
@@ -170,7 +141,8 @@ function PartnerUpdateTemplate({ postId }: { postId: string }) {
   }
 
   // 글 작성 버튼 클릭 핸들러
-  const handleUpdateClick = () => {
+  const handleUpdateClick = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
+    e.preventDefault();
     if (validation()) {
       const dataToInsert = {
         id: postId,
