+++ src/components/partner/userFeedback/UserFeedback.tsx
@@ -4,7 +4,7 @@ import { useEffect, useState } from 'react';
 import { FiEdit, FiTrash2 } from 'react-icons/fi';
 import { RiBookmarkFill, RiBookmarkLine } from 'react-icons/ri';
 import { useNavigate } from 'react-router';
-import { addBookmark, bookmarkCheck, deletePartnerPost, removeBookMark } from '../../../api/supabase/partner';
+import { addBookmark, bookmarkCheck, deletePartnerPost, getApplicantList, getConfirmedApplicantList, removeBookMark } from '../../../api/supabase/partner';
 import { Tables } from '../../../api/supabase/supabase';
 import Messages from '../../../assets/imgs/partner/Messages.png';
 import defaultProfileImage from '../../../assets/imgs/users/default_profile_img.png';
@@ -25,6 +25,9 @@ const UserFeedback = ({ partnerPostData }: Props) => {
   const [bookMark, setBookMark] = useState(false);
   const navigate = useNavigate();
   const queryClient = useQueryClient();
+
+  const [applicantList, setApplicantList] = useState<Tables<'applicants'>[]>([]);
+  const [confirmedApplicantList, setConfirmedApplicantList] = useState<Tables<'applicants'>[]>([]);
 
   const storagaUrl = process.env.REACT_APP_SUPABASE_STORAGE_URL;
 
@@ -63,6 +66,30 @@ const UserFeedback = ({ partnerPostData }: Props) => {
     Alert({ title: '복사 완료!' });
   };
 
+  // 동행 모집이 시작되면 글 수정 페이지로 못 넘거가도록
+  useEffect(() => {
+    const fetchApplicant = async () => {
+      if (!id) return;
+      const { data, error } = await getApplicantList(id);
+      if (error || !data) {
+        setApplicantList([]);
+      } else {
+        setApplicantList(data);
+      }
+    };
+    fetchApplicant();
+
+    const fetchConfirmedPartnerList = async () => {
+      if (id) {
+        const response = await getConfirmedApplicantList(id!);
+        if (response.data !== null) {
+          setConfirmedApplicantList(response.data);
+        }
+      }
+    };
+    fetchConfirmedPartnerList();
+  }, [id]);
+
   const session = useSessionStore((state) => state.session);
   const logInUserId = session?.user.id;
 
@@ -78,7 +105,20 @@ const UserFeedback = ({ partnerPostData }: Props) => {
 
   const isPostUser = () => logInUserId === writerId;
 
-  //글삭제
+  // 글 수정
+  const handleEdit = () => {
+    if (confirmedApplicantList.length === partnerPostData.numOfPeople) {
+      AlertError({ title: '수정이 불가능 합니다.', text: '동행 모집이 완료되었습니다.' });
+      return false;
+    }
+    if (applicantList.length >= 1 || confirmedApplicantList.length >= 1) {
+      AlertError({ title: '수정이 불가능 합니다.', text: '동행 신청이 시작되었습니다.' });
+      return false;
+    }
+    navigate(`/partner/write/${id}`);
+  };
+
+  // 글 삭제
   const handleDelBtn = async (id: string) => {
     const isConfirmed = await ConfirmDelete('해당 동행 글이 삭제되었습니다.');
     if (!isConfirmed) {
@@ -86,6 +126,7 @@ const UserFeedback = ({ partnerPostData }: Props) => {
     }
     mutation.mutate({ postId: id });
   };
+
   return (
     <St.UserFeedbackBox>
       <St.UserProfileBox>
@@ -108,7 +149,7 @@ const UserFeedback = ({ partnerPostData }: Props) => {
 
         {isPostUser() ? (
           <>
-            <button>{<FiEdit className="lineIcon" onClick={() => navigate(`/partner/write/${id}`)} />}</button>
+            <button>{<FiEdit className="lineIcon" onClick={handleEdit} />}</button>
             <button>{<FiTrash2 className="lineIcon" onClick={() => handleDelBtn(id)} />}</button>
           </>
         ) : (
